mfpt for stair dealt with
